+++ Tests/UIKitTests/UITextFieldExtensionsTests.swift
@@ -135,16 +135,32 @@ final class UITextFieldExtensionsTests: XCTestCase {
         XCTAssertEqual(textfield.leftView?.frame.width, 40)
     }
 
-    func testAddPaddingImageIcon() {
+    func testAddPaddingRight() {
+        let textfield = UITextField()
+        textfield.frame = CGRect(x: 0, y: 0, width: 100, height: 30)
+        textfield.addPaddingRight(40)
+        XCTAssertEqual(textfield.rightView?.frame.width, 40)
+    }
+
+    func testAddPaddingImageLeftIcon() {
         let textfield = UITextField()
         textfield.frame = CGRect(x: 0, y: 0, width: 100, height: 44)
 
-        let bundle = Bundle.init(for: UIImageExtensionsTests.self)
+        let bundle = Bundle(for: UIImageExtensionsTests.self)
         let image = UIImage(named: "TestImage", in: bundle, compatibleWith: nil)!
         textfield.addPaddingLeftIcon(image, padding: 5)
         XCTAssertEqual(textfield.leftView?.frame.width, image.size.width + 5)
     }
 
+    func testAddPaddingImageRightIcon() {
+        let textfield = UITextField()
+        textfield.frame = CGRect(x: 0, y: 0, width: 100, height: 44)
+
+        let bundle = Bundle(for: UIImageExtensionsTests.self)
+        let image = UIImage(named: "TestImage", in: bundle, compatibleWith: nil)!
+        textfield.addPaddingRightIcon(image, padding: 5)
+        XCTAssertEqual(textfield.rightView?.frame.width, image.size.width + 5)
+    }
 }
 
 #endif